The name of this source file (`Enumerable.rb`) should use snake_case.
